requirements.txt
requests
python-dotenv
anthropic>=0.19.0
